Split up from the controller
